chatbot-backend/requirements.txt
flask
flask-cors
gunicorn
tensorflow-cpu
numpy
protobuf